+++ modules/table/src/test/java/org/apache/ignite/internal/table/distributed/raft/snapshot/FullStateTransferIndexChooserTest.java
@@ -56,18 +56,16 @@ import org.apache.ignite.internal.hlc.HybridClockImpl;
 import org.apache.ignite.internal.hlc.HybridTimestamp;
 import org.apache.ignite.internal.table.TableTestUtils;
 import org.apache.ignite.internal.table.distributed.LowWatermark;
+import org.apache.ignite.internal.table.distributed.TestLowWatermarkImpl;
 import org.apache.ignite.internal.testframework.BaseIgniteAbstractTest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 /** For {@link FullStateTransferIndexChooser} testing. */
-@ExtendWith(MockitoExtension.class)
+// TODO: IGNITE-21514 поправить тесты и написать новые
 public class FullStateTransferIndexChooserTest extends BaseIgniteAbstractTest {
     private static final String REGISTERED_INDEX_NAME = INDEX_NAME + "_" + REGISTERED;
 
@@ -83,15 +81,12 @@ public class FullStateTransferIndexChooserTest extends BaseIgniteAbstractTest {
 
     private CatalogManager catalogManager;
 
-    @Mock
-    private LowWatermark lowWatermark;
+    private final LowWatermark lowWatermark = new TestLowWatermarkImpl();
 
     private FullStateTransferIndexChooser indexChooser;
 
     @BeforeEach
     void setUp() {
-        // TODO: IGNITE-21514 поправить тесты и написать новые
-
         catalogManager = CatalogTestUtils.createTestCatalogManager("test", clock);
 
         indexChooser = new FullStateTransferIndexChooser(catalogManager, lowWatermark);
